% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paramCalcs.R
\name{alphaCalc}
\alias{alphaCalc}
\title{Calculate the source to indoor air attenuation factor (alpha)}
\usage{
alphaCalc(Aparam, Bparam, Cparam, Qsoil_Qb, foundation_type, source_medium)
}
\arguments{
\item{Aparam}{The Johnson and Ettinger model equation "A" parameter value
(dimensionless)}

\item{Bparam}{The Johnson and Ettinger model equation "B" parameter value
(dimensionless)}

\item{Cparam}{The Johnson and Ettinger model equation "C" parameter value
(dimensionless)}

\item{Qsoil_Qb}{The ratio of the average vapor flow rate into the building
divided by the building ventilation rate (dimensionless)}

\item{foundation_type}{The foundation type. Choose from the following:
\itemize{
\item "Slab-grade" for a slab-on-grade foundation
\item "Basement-slab" for a basement with slab foundation
\item "Basement-dirt" for a basement with dirt floor foundation
\item "Crawlspace-slab" for a crawlspace with slab foundation
\item "Crawlspace-dirt" for a crawlspace with dirt floor foundation
}}

\item{source_medium}{The source medium. Choose from "Groundwater", "Subslab
Soil Gas", or "Exterior Soil Gas"}
}
\value{
The source to indoor air attenuation factor
(dimensionless)
}
\description{
Calculates the source to indoor air attenuation factor
(dimensionless). For groundwater, the source concentration is the soil gas
concentration in equilibrium with the groundwater concentration.
}
\examples{
alphaCalc(Aparam = 4.8E-6,
Bparam = 6.8E2,
Cparam = 3.0E-3,
Qsoil_Qb = 0.0030,
foundation_type = "Slab-grade",
source_medium = "Groundwater")

}
