% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paramCalcs.R
\name{AparamCalc}
\alias{AparamCalc}
\title{Calculate the "A" parameter in the Johnson and Ettinger model equation}
\usage{
AparamCalc(DeffT, Abf, Lb, Qb, Ls)
}
\arguments{
\item{DeffT}{Effective diffusion coefficient from the source to the building
foundation in square centimeters per second (\ifelse{html}{\out{cm<sup>2</sup>}}{\eqn{cm^2/sec}})}

\item{Abf}{Enclosed space floor area in square meters (\ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^2}})}

\item{Lb}{Depth below grade to base of foundation in meters (\ifelse{html}{\out{m}}{\eqn{m}})}

\item{Qb}{Building ventilation rate in cubic meters per hour (\ifelse{html}{\out{m<sup>3</sup>/hr}}{\eqn{m^3/hr}})}

\item{Ls}{Depth below grade to water table (\ifelse{html}{\out{m}}{\eqn{m}})}
}
\value{
The Johnson and Ettinger model equation "A" parameter value (dimensionless)
}
\description{
Calculates the "A" parameter value in the Johnson and Ettinger
model equation, which represents diffusive transport from the source to
a building with a dirt floor foundation (dimensionless)
}
\examples{
AparamCalc(DeffT = 0.000133929,
Abf = 150.00,
Lb = 0.10,
Qb = 164.70,
Ls = 9.54)

}
