% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_boringLogSetup.R
\name{boringLogSetup}
\alias{boringLogSetup}
\title{Assign soil strata log properties}
\usage{
boringLogSetup(boring_log_dfx)
}
\arguments{
\item{boring_log_dfx}{Dataframe of soil strata log information containing the
following required fields:
\itemize{
\item SoilType: One of 12 USDA SCS soil types. Supported values are "Clay",
"Clay Loam", "Loam", "Loamy Sand", "Sand", "Sandy Clay", "Sandy Clay Loam",
"Sandy Loam", "Silt", "Silt Loam", "Silty Clay", and "Silty Clay Loam".
\item LayerOrder: Layer order from ground surface. Layer 1 is the surface layer,
layer 2 is beneath layer 1, and so on.
\item Thickness: Layer thickness. The
package assumes Thickness values are entered in meters when this function is
used.
}

Note that the LogID and Units fields associated with soil strata data in the
import template are not used within this function}
}
\value{
The same data frame that was entered as a function input, but with
the following fields added:
\itemize{
\item SCS: ID number associated with the layer's SCS soil type in the data table
of standard soil type properties
\item nS: Layer total porosity (dimensionless)
\item nwS: Layer water-filled porosity (dimensionless)
\item rhoS: Layer bulk
density (grams per cubic centimeter, or \ifelse{html}{\out{g/cm<sup>2</sup>}}{\eqn{g/cm^2}})
\item ncz: Capillary zone total
porosity (dimensionless)
\item nwcz: Capillary zone water-filled porosity
(dimensionless)
\item hcz: Capillary fringe height (meters, or m)
\item upper_depth: Layer upper depth, in the same units as the Thickness field values
\item lower_depth: Layer lower depth, in the same units as the Thickness field
values
}
}
\description{
Performs basic setup steps required by \code{runJE()} for
soil strata log data. Assigns United States Department of Agriculture (USDA)
Soil Conservation Service (SCS) soil type properties, sorts the soil
strata layers by LayerOrder, and identifies the upper and lower depth of
each layer.
}
\examples{
# Get example soil strata log data
soil_strata_log_dfx <- det_jem_sim_example_data[[4]]

# Run the boring log setup function
boringLogSetup(soil_strata_log_dfx)

}
