% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paramCalcs.R
\name{BparamCalc}
\alias{BparamCalc}
\title{Calculate the "B" parameter in the Johnson and Ettinger model equation}
\usage{
BparamCalc(Qsoil_Qb, Qb, Lf, DeffBF, eta, Abf, Lb)
}
\arguments{
\item{Qsoil_Qb}{The ratio of the average vapor flow rate into the building
divided by the building ventilation rate (dimensionless)}

\item{Qb}{The building ventilation rate in cubic meters per hour (\ifelse{html}{\out{m<sup>3</sup>/hr}}{\eqn{m^3/hr}})}

\item{Lf}{The foundation thickness in meters (\ifelse{html}{\out{m}}{\eqn{m}})}

\item{DeffBF}{The effective diffusion coefficient in the soil layer
containing the building foundation in square centimeters per second
(\ifelse{html}{\out{cm<sup>2</sup>/sec}}{\eqn{cm^2/sec}})}

\item{eta}{The fraction of the foundation area with cracks (dimensionless)}

\item{Abf}{The enclosed space floor area in square meters (\ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^2}})}

\item{Lb}{The depth below grade to the base of the foundation (\ifelse{html}{\out{m}}{\eqn{m}})}
}
\value{
The Johnson and Ettinger model equation "B" parameter value (dimensionless)
}
\description{
Calculates the "B" parameter in the Johnson and Ettinger model
equation, which represents the Peclet Number (Pe) for contaminant transport
through the building foundation, or the ratio of advective to diffusive
transport (dimensionless)
}
\examples{
BparamCalc(Qsoil_Qb = 0.0030,
Qb = 164.70,
Lf = 0.10,
DeffBF = 0.00130364,
eta = 0.001,
Abf = 150.00,
Lb = 0.10)

}
