% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_checkModelInputs.R
\name{checkModelInputs}
\alias{checkModelInputs}
\title{Confirm that all inputs to \code{runJE()} are formatted correctly}
\usage{
checkModelInputs(
  contaminant_data,
  building_data,
  vadose_zone_data,
  strata_logs_data,
  settings_data
)
}
\arguments{
\item{contaminant_data}{A data frame or \code{JEMParamDist} object with input
contaminant data}

\item{building_data}{A data frame or list of \code{JEMParamDist} objects with input
building data}

\item{vadose_zone_data}{A data frame or list of \code{JEMParamDist} objects with
input vadose zone data}

\item{strata_logs_data}{A data frame with input data for the strata log or
logs that characterize the subsurface soils beneath the building}

\item{settings_data}{A data frame containing input simulation settings}
}
\value{
The \code{checkModelInputs()} function will throw an error anytime the
input parameters would prevent \code{runJE()} from executing successfully. In
addition, it will also throw warnings if the parameter values are outside
of typical ranges, and in other instances. The value returned by
\code{checkModelInputs()} is a string that identifies the thrown warning
messages. If there are not any warnings, the string returned will state "No
errors or warnings in model inputs".
}
\description{
Checks that all required inputs are present to run the Johnson
and Ettinger model using \code{runJE()}, and that all inputs are
valid and in the proper format. The function input parameters are the
same as those used by \code{runJE()}.
}
\examples{
  #Get example inputs to runJE()
  contaminant_data <- det_jem_sim_example_data[[1]]
  building_data <- det_jem_sim_example_data[[2]]
  vadose_zone_data <- det_jem_sim_example_data[[3]]
  strata_logs_data <- det_jem_sim_example_data[[4]]
  settings_data <- det_jem_sim_example_data[[5]]

  #Run checkModelInputs with no warnings issued
  checkModelInputs(contaminant_data, building_data, vadose_zone_data,
  strata_logs_data, settings_data)

  #Set the building air exchange rate to a large value so that
  #checkModelInputs issues a warning
  building_data$ach <- 5

  #Run checkModelInputs and generate the warning message
  checkModelInputs(contaminant_data, building_data, vadose_zone_data,
  strata_logs_data, settings_data)

}
