% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paramCalcs.R
\name{Cia_ppbCalc}
\alias{Cia_ppbCalc}
\title{Calculate the modeled indoor air concentration in units of parts per
billion (Cia_ppb)}
\usage{
Cia_ppbCalc(Cs, alpha, MW)
}
\arguments{
\item{Cs}{The source vapor concentration in micrograms per cubic meter (\ifelse{html}{\out{&microg/m<sup>3</sup>}}{\eqn{{\mu}g/m^3}})}

\item{alpha}{The source to indoor air attenuation factor (dimensionless)}

\item{MW}{The chemical molecular weight in grams per mole (\ifelse{html}{\out{g/mol}}{\eqn{g/mol}})}
}
\value{
Cia_ppb
}
\description{
Calculates the modeled indoor air concentration in parts per
billion (ppb) assuming standard temperature and pressure (25 degrees
Celsius and 1 atmosphere)
}
\examples{
Cia_ppbCalc(Cs = 7238.379804,
alpha = 4.79586E-06,
MW = 165.83)

}
