% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paramCalcs.R
\name{CiaCalc}
\alias{CiaCalc}
\title{Calculate the modeled indoor air concentration}
\usage{
CiaCalc(Cs, alpha)
}
\arguments{
\item{Cs}{The source vapor concentration (\ifelse{html}{\out{&microg/m<sup>3</sup>}}{\eqn{{\mu}g/m^3}})}

\item{alpha}{The source to indoor air attenuation factor (dimensionless)}
}
\value{
The modeled indoor air concentration (\ifelse{html}{\out{&microg/m<sup>3</sup>}}{\eqn{{\mu}g/m^3}})
}
\description{
Calculates the modeled indoor air concentration due to vapor
intrusion in micrograms per cubic meter (\ifelse{html}{\out{&microg/m<sup>3</sup>}}{\eqn{{\mu}g/m^3}})
}
\examples{
CiaCalc(Cs = 7238.379804,
alpha = 4.79586E-06)

}
