% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_ConcDataFig.R
\name{ConcDataFig}
\alias{ConcDataFig}
\title{Parent class for measured and modeled data figures}
\description{
An R6 parent class used to create \code{ModDataFig} and \code{MeasDataFig}
objects. The class contains common properties and methods needed to
generate figures with the two child classes. See the following vignette
for more information.

\code{vignette("ModelResultFigures", package = "vapintr")}
}
\examples{
#Run the JEM using the example stochastic data
jem_results <- runTemplateData(stoc_jem_sim_example_data, "Trichloroethylene")

#Create a new output figure from the JEM Monte Carlo results
model_result_figure <- ModelResultFig$new(jem_results[[1]]$JEMResults,
"Trichloroethylene", "79-01-6", "ug/m3", "groundwater")

#The object stored in the model_result_figure$mod_data_fig attribute is a
#ModDataFig object, which uses the ConcDataFig as its parent class
model_result_figure$mod_data_fig

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data_plot}}{Primary ggplot of concentration data associated with the object}

\item{\code{title}}{Figure title}

\item{\code{x_title}}{Title label for the x-axis}

\item{\code{x_min}}{Minimum value of data on the x-axis}

\item{\code{x_max}}{Maximum value of data on the x-axis}

\item{\code{y_title}}{Title label for the y-axis}

\item{\code{y_min}}{Minimum value of data on the y-axis}

\item{\code{y_max}}{Maximum value of data on the y-axis}

\item{\code{label_units}}{Units formatted to appear in the y-axis concentration label}

\item{\code{factor_indices}}{Indices associated with the standard data type factors used in the figure}

\item{\code{factor_names}}{Names associated with the standard data type factors used in the figure}

\item{\code{factor_line_formats}}{Line formats associated with the standard data type factors used in the figure}

\item{\code{factor_shape_formats}}{Shape formats associated with the standard data type factors used in the figure}

\item{\code{factor_color_formats}}{Color formats associated with the standard data type factors used in the figure}

\item{\code{format_constants}}{Formatting constants for generating figures}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConcDataFig-new}{\code{ConcDataFig$new()}}
\item \href{#method-ConcDataFig-getStandardFormatConstants}{\code{ConcDataFig$getStandardFormatConstants()}}
\item \href{#method-ConcDataFig-getStandardFactorIndices}{\code{ConcDataFig$getStandardFactorIndices()}}
\item \href{#method-ConcDataFig-getStandardFactorLineFormats}{\code{ConcDataFig$getStandardFactorLineFormats()}}
\item \href{#method-ConcDataFig-getStandardFactorShapeFormats}{\code{ConcDataFig$getStandardFactorShapeFormats()}}
\item \href{#method-ConcDataFig-getStandardFactorNames}{\code{ConcDataFig$getStandardFactorNames()}}
\item \href{#method-ConcDataFig-getStandardFactorColorFormats}{\code{ConcDataFig$getStandardFactorColorFormats()}}
\item \href{#method-ConcDataFig-addPlotLabels}{\code{ConcDataFig$addPlotLabels()}}
\item \href{#method-ConcDataFig-clone}{\code{ConcDataFig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConcDataFig-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConcDataFig-new}{}}}
\subsection{Method \code{new()}}{
Function to create a new \code{ConcDataFig} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcDataFig$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConcDataFig-getStandardFormatConstants"></a>}}
\if{latex}{\out{\hypertarget{method-ConcDataFig-getStandardFormatConstants}{}}}
\subsection{Method \code{getStandardFormatConstants()}}{
Gets a named list of standard figure format constants
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcDataFig$getStandardFormatConstants()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConcDataFig-getStandardFactorIndices"></a>}}
\if{latex}{\out{\hypertarget{method-ConcDataFig-getStandardFactorIndices}{}}}
\subsection{Method \code{getStandardFactorIndices()}}{
Gets a named list of indices associated with the figure factors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcDataFig$getStandardFactorIndices()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConcDataFig-getStandardFactorLineFormats"></a>}}
\if{latex}{\out{\hypertarget{method-ConcDataFig-getStandardFactorLineFormats}{}}}
\subsection{Method \code{getStandardFactorLineFormats()}}{
Gets a vector of standard line formats associated with the figure factors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcDataFig$getStandardFactorLineFormats()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConcDataFig-getStandardFactorShapeFormats"></a>}}
\if{latex}{\out{\hypertarget{method-ConcDataFig-getStandardFactorShapeFormats}{}}}
\subsection{Method \code{getStandardFactorShapeFormats()}}{
Gets a vector of standard shape formats associated with the figure factors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcDataFig$getStandardFactorShapeFormats()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConcDataFig-getStandardFactorNames"></a>}}
\if{latex}{\out{\hypertarget{method-ConcDataFig-getStandardFactorNames}{}}}
\subsection{Method \code{getStandardFactorNames()}}{
Gets a vector of names associated with the figure factors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcDataFig$getStandardFactorNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConcDataFig-getStandardFactorColorFormats"></a>}}
\if{latex}{\out{\hypertarget{method-ConcDataFig-getStandardFactorColorFormats}{}}}
\subsection{Method \code{getStandardFactorColorFormats()}}{
Gets a vector of standard colors associated with the figure factors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcDataFig$getStandardFactorColorFormats()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConcDataFig-addPlotLabels"></a>}}
\if{latex}{\out{\hypertarget{method-ConcDataFig-addPlotLabels}{}}}
\subsection{Method \code{addPlotLabels()}}{
Adds labels to the plot title, x-axis, and y-axis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcDataFig$addPlotLabels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConcDataFig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConcDataFig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcDataFig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
