% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_concUnitConvs.R
\name{concUnitConvs}
\alias{concUnitConvs}
\title{Convert concentrations to new units}
\usage{
concUnitConvs(inputConc, inputUnits, outputUnits, medium, Ts = NA, MW = NA)
}
\arguments{
\item{inputConc}{Concentration to be converted to a new set of units}

\item{inputUnits}{Units associated with the input concentration. Must be one
of the following: "mg/L", "mg/m3", "ug/L", "ug/m3", "ppm", or "ppb".}

\item{outputUnits}{Units associated with the output concentration. Must be
one of the following: "mg/L", "mg/m3", "ug/L", "ug/m3", "ppm", or "ppb".}

\item{medium}{Medium for the input concentration. Can be either "Water" or
"Air".}

\item{Ts}{Temperature of the medium. Required for air unit conversions from a
"parts per" system to a "mass/volume" system and vice-versa. Assumed to be
input in degrees Celsius.}

\item{MW}{Chemical molecular weight. Required for air unit conversions from a
"parts per" system to a "mass/volume" system and vice-versa. Assumed to be
input in grams/mole.}
}
\value{
Returns the input concentration converted to the output units
}
\description{
Performs unit conversions for units commonly
associated with contaminant concentrations in air and water. Supported units are:
\itemize{
\item milligrams per liter (mg/L)
\item milligrams per cubic meter (mg/m3)
\item micrograms per liter (ug/L)
\item micrograms per cubic meter (ug/m3)
\item parts per million (ppm)
\item parts per billion (ppb)
}
}
\examples{
#Convert a water concentration from ppm to ug/L
concUnitConvs(1, "ppm", "ug/L", "Water")

#Convert an air concentration from ppb to ug/m3
concUnitConvs(1, "ppb", "ug/m3", "Air", Ts = 25, MW = 165.83)

}
