% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paramCalcs.R
\name{CparamCalc}
\alias{CparamCalc}
\title{Calculate the "C" parameter in the Johnson and Ettinger model equation}
\usage{
CparamCalc(Qsoil_Qb)
}
\arguments{
\item{Qsoil_Qb}{The ratio of the average vapor flow rate into the building
divided by the building ventilation rate (dimensionless)}
}
\value{
The Johnson and Ettinger model equation "C" parameter value (dimensionless)
}
\description{
Calculates the "C" parameter value in the Johnson and Ettinger
model equation, which represents convective transport from subslab
to the building (dimensionless)
}
\examples{
CparamCalc(Qsoil_Qb = 0.0030)

}
