% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paramCalcs.R
\name{CsCalc}
\alias{CsCalc}
\title{Calculate the source vapor concentration (Cs)}
\usage{
CsCalc(Cmedium, Hs, source_medium)
}
\arguments{
\item{Cmedium}{Source concentration in units of micrograms per liter (\ifelse{html}{\out{&microg/L}}{\eqn{{\mu}g/L}})
for groundwater or  \ifelse{html}{\out{&microg/m<sup>3</sup>}}{\eqn{{\mu}g/m^3}} for subslab soil gas and exterior soil gas}

\item{Hs}{Henry's Law constant at system temperature (dimensionless)}

\item{source_medium}{The source medium. Choose from "Groundwater", "Subslab
Soil Gas", or "Exterior Soil Gas"}
}
\value{
The source vapor concentration in \ifelse{html}{\out{&microg/m<sup>3</sup>}}{\eqn{{\mu}g/m^3}}
}
\description{
Calculates the source vapor concentration in micrograms per
cubic meter ( \ifelse{html}{\out{&microg/m<sup>3</sup>}}{\eqn{{\mu}g/m^3}})
}
\examples{
CsCalc(Cmedium = 10,
 Hs = 0.724,
 source_medium = "Groundwater")

}
