% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paramCalcs.R
\name{CssCalc}
\alias{CssCalc}
\title{Calculate the subslab vapor concentration (Css)}
\usage{
CssCalc(Cia, Qb, Qsoil, foundation_type)
}
\arguments{
\item{Cia}{The modeled indoor air concentration due to vapor intrusion (\ifelse{html}{\out{&microg/m<sup>3</sup>}}{\eqn{{\mu}g/m^3}})}

\item{Qb}{The building ventilation rate in cubic meters per hour (\ifelse{html}{\out{m<sup>3</sup>/hr}}{\eqn{m^3/hr}})}

\item{Qsoil}{The average vapor flow rate into the building (\ifelse{html}{\out{m<sup>3</sup>/hr}}{\eqn{m^3/hr}})}

\item{foundation_type}{The foundation type. Choose from the following:
\itemize{
\item "Slab-grade" for a slab-on-grade foundation
\item "Basement-slab" for a basement with slab foundation
\item "Basement-dirt" for a basement with dirt floor foundation
\item "Crawlspace-slab" for a crawlspace with slab foundation
\item "Crawlspace-dirt" for a crawlspace with dirt floor foundation
}}
}
\value{
The subslab vapor concentration (\ifelse{html}{\out{&microg/m<sup>3</sup>}}{\eqn{{\mu}g/m^3}}). The function returns \code{NA} for
buildings with dirt floor foundations.
}
\description{
Calculates the subslab vapor concentration in units of
micrograms per cubic meter (\ifelse{html}{\out{&microg/m<sup>3</sup>}}{\eqn{{\mu}g/m^3}})
}
\examples{
CssCalc(Cia = 0.034714258,
Qb = 164.70,
Qsoil = 0.4941,
foundation_type = "Slab-grade")

}
