% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paramCalcs.R
\name{DeffCalc}
\alias{DeffCalc}
\title{Calculate the effective diffusion coefficient (Deff)}
\usage{
DeffCalc(Dair, Dwater, nS, nwS, Hs)
}
\arguments{
\item{Dair}{Diffusion coefficient in air (\ifelse{html}{\out{cm<sup>2</sup>/sec}}{\eqn{cm^2/sec}})}

\item{Dwater}{Diffusion coefficient in water (\ifelse{html}{\out{cm<sup>2</sup>/sec}}{\eqn{cm^2/sec}})}

\item{nS}{Stratum total porosity (unitless)}

\item{nwS}{Stratum water filled porosity (unitless)}

\item{Hs}{Henry's Law constant at system temperature (unitless)}
}
\value{
The chemical effective diffusion coefficient (\ifelse{html}{\out{cm<sup>2</sup>/sec}}{\eqn{cm^2/sec}})
}
\description{
Calculates the chemical effective diffusion coefficient in one
soil layer, in square centimeters per second (\ifelse{html}{\out{cm<sup>2</sup>/sec}}{\eqn{cm^2/sec}})
}
\examples{
DeffCalc(Dair = 0.0504664,
    Dwater = 9.4551E-06,
    nS = 0.385,
    nwS = 0.197,
    Hs = 0.72383798)

}
