% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_DeffMultiLayerCalc.R
\name{DeffMultiLayerCalc}
\alias{DeffMultiLayerCalc}
\title{Calculate the effective diffusivity across multiple soil layers}
\usage{
DeffMultiLayerCalc(layer_dfx, Dair, Dwater, Hs, is_capillary_zone)
}
\arguments{
\item{layer_dfx}{A data frame of soil layers. Required fields for this data
frame are assigned in \code{boringLogSetup()}.}

\item{Dair}{Contaminant diffusion coefficient in air (square centimeters per
second, or \ifelse{html}{\out{cm<sup>2</sup>/sec}}{\eqn{cm^2/sec}})}

\item{Dwater}{Contaminant diffusion coefficient in water (\ifelse{html}{\out{cm<sup>2</sup>/sec}}{\eqn{cm^2/sec}})}

\item{Hs}{Henry's Law constant at system temperature (dimensionless)}

\item{is_capillary_zone}{A true/false flag identifying whether the layers are
in the capillary zone}
}
\value{
The effective diffusivity across all the entered layers in
\ifelse{html}{\out{cm<sup>2</sup>/sec}}{\eqn{cm^2/sec}}.
}
\description{
Calculates a weighted effective diffusivity for a contaminant
across multiple soil layers. The weights are determined by the thickness of
each layer.
}
\examples{
#Get example soil strata log data
soil_strata_log_data <- det_jem_sim_example_data[[4]]

#Assign properties to each soil layer
soil_strata_log_data <- boringLogSetup(soil_strata_log_data)

#Set example values for chemical properties (tetrachloroethylene)
Dair <- 0.0504664 #cm2/sec
Dwater <- 9.4551e-06 #cm2/sec
Hs <- 0.0177 #dimensionless

#Calculate the effective diffusivity
DeffMultiLayerCalc(soil_strata_log_data, Dair, Dwater, Hs, FALSE)

}
