% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{det_jem_sim_example_data}
\alias{det_jem_sim_example_data}
\title{Deterministic simulation data for running function examples}
\format{
An object of class \code{list} of length 6.
}
\usage{
det_jem_sim_example_data
}
\description{
A list of 6 elements used for deterministic simulation function
examples. The data are the same as those in the import file described
in the \verb{Deterministic Simulations} vignette. The first 5 output elements
correspond with the following inputs to \code{runJE()}. The 6th output is an
empty data frame of reference air concentration data.
\enumerate{
\item contaminant_data
\item building_data
\item vadose_zone_data
\item strata_logs_data
\item settings_data
}
}
\keyword{datasets}
