% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_getBuildingDefaults.R
\name{getBuildingDefaults}
\alias{getBuildingDefaults}
\title{Get default values for building parameters}
\usage{
getBuildingDefaults(building_setting, foundation_type, simulation_type)
}
\arguments{
\item{building_setting}{The building type. Choose from "Residential" or
"Commercial".}

\item{foundation_type}{The foundation type. Choose from the following:
\itemize{
\item "Slab-grade" for a slab-on-grade foundation
\item "Basement-slab" for a basement with slab foundation
\item "Basement-dirt" for a basement with dirt floor foundation
\item "Crawlspace-slab" for a crawlspace with slab foundation
\item "Crawlspace-dirt" for a crawlspace with dirt floor foundation
}}

\item{simulation_type}{The simulation type. Choose from "DET" for a
deterministic simulation or "MC" for a stochastic simulation.}
}
\value{
A data frame of building parameter values. For deterministic
simulations, the data frame will have one record per parameter value. For
stochastic simulations, the data frame will have multiple records per
parameter value, with each record identifying the distribution type and a
distribution parameter value.
}
\description{
Gets default parameter values for a specified building and
foundation type for use in deterministic and stochastic simulations. The
default parameter values come from standard parameter values included in
version 6.0 of USEPA's Johnson and Ettinger model spreadsheet. See the
following vignettes for more information.

\code{vignette("DeterministicSimulations", package = "vapintr")}

\code{vignette("StochasticSimulations", package = "vapintr")}
}
\examples{
getBuildingDefaults("Residential", "Basement-dirt", "DET")
getBuildingDefaults("Commercial", "Slab-grade", "MC")

}
