% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_getChemProperties.R
\name{getChemicalProperties}
\alias{getChemicalProperties}
\title{Get standard properties for a specific chemical}
\usage{
getChemicalProperties(Chemical = NA, CAS = NA)
}
\arguments{
\item{Chemical}{Chemical name. Enter a value for this parameter or for the CAS
parameter to identify the chemical. If values are entered for both
parameters, they must correspond to the same chemical. Only the chemical
names in \code{chem_data$Chemical} are supported.}

\item{CAS}{Chemical Abstract Service Registry Number. Enter a value for this
parameter or for the Chemical parameter to identify the chemical. If values are
entered for both parameters, they must correspond to the same chemical.
Only the CAS numbers in \code{chem_data$CAS} are supported.}
}
\value{
A data frame of standard chemical properties for the specified
chemical.
}
\description{
Gets a data frame of standard chemical properties for a specific
chemical. The standard chemical properties included in vapintr were obtained
from version 6.0 of USEPA's Johnson and Ettinger model spreadsheet.
}
\examples{
getChemicalProperties(Chemical = "Tetrachloroethylene")
getChemicalProperties(CAS = "127-18-4")

}
