% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_getLegend.R
\name{getLegend}
\alias{getLegend}
\title{Get legend from ggplot2 ver 3.5}
\usage{
getLegend(plot, legend_number = 1)
}
\arguments{
\item{plot}{ggplot2 object to have legend extracted}

\item{legend_number}{which plot to extract}
}
\description{
Calculates Hs using default chemical properties at the
subsurface system temperature. Code is from https://github.com/wilkelab/cowplot/issues/202
}
