% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_getSCSTypeNumber.R
\name{getSCSTypeNumber}
\alias{getSCSTypeNumber}
\title{Get a soil type index for the soil type property reference table}
\usage{
getSCSTypeNumber(scs_type)
}
\arguments{
\item{scs_type}{The SCS soil type.}
}
\value{
The numeric SCS soil type index
}
\description{
Converts a United States Department of Agriculture Soil
Conservation Service (SCS) soil type into a numeric index for looking up
properties in the SCS reference table (\code{SCS_soil}).
}
\examples{
getSCSTypeNumber("Silty Clay")

}
