% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hczCalc.R
\name{hczCalc}
\alias{hczCalc}
\title{Calculate the height of the capillary zone}
\usage{
hczCalc(boring_log_dfx, Ls)
}
\arguments{
\item{boring_log_dfx}{A data frame of soil strata data beneath the building.
The required fields for this data frame are added to data imported from the
vapintr data import template in \code{boringLogSetup()}.}

\item{Ls}{The depth to the water table (\ifelse{html}{\out{m}}{\eqn{m}}).}
}
\value{
The height of the capillary zone (\ifelse{html}{\out{m}}{\eqn{m}}).
}
\description{
Calculates the height of the capillary zone in meters (m).
The function implements the following algorithm, which accounts for
capillary rise across multiple soil layers.
\enumerate{
\item Beginning at the water table, first identify the capillary rise
associated with the soil type of each layer above the water table. The
first layer is always included in the capillary zone. If the thickness of
the first layer is greater than the capillary rise of the first layer, then
stop.
\item If the capillary rise of the second layer is greater than the distance
from the base of the layer to the water table, the second layer also
contributes to capillary rise and the capillary rise is that of the second
layer. If the capillary rise of the second layer does not meet this
condition, the layer may still be part of the capillary zone depending on
the capillary rise in the third layer (if present). If the total thickness
of the first two layers is greater than the overall capillary rise, then
stop.
\item If a third layer is present and its capillary rise is greater than the
distance from the base of the layer to the water table, then the capillary
rise is that of the third layer. If the total thickness of the first three
layers is greater than the overall capillary rise, then stop.
\item Continue testing each layer above the water table until a stop point is
reached or until the maximum capillary rise (192.31 centimeters for silty
clay) is achieved.
}
}
\examples{
#Get example soil strata log data
soil_strata_log_data <- det_jem_sim_example_data[[4]]

#Assign properties to each soil layer
soil_strata_log_data <- boringLogSetup(soil_strata_log_data)

#Get the height of the capillary zone when the water table is 10 meters deep
#for the example soil strata log data
hczCalc(soil_strata_log_data, 10)
}
