% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paramCalcs.R
\name{HrCalc}
\alias{HrCalc}
\title{Calculate the dimensionless Henry's Law constant (Hr)}
\usage{
HrCalc(Hc)
}
\arguments{
\item{Hc}{Henry's law constant at 25 \ifelse{html}{\out{&degC}}{\eqn{^{\circ}C}} (\ifelse{html}{\out{atm-m<sup>3</sup>/L}}{\eqn{atm-m^3/mol}})}
}
\value{
Dimensionless Henry's Law constant (dimensionless)
}
\description{
Calculates the dimensionless Henry's Law constant from a
reference value of Henry's Law constant in units of atmospheres-cubic
meters per mole (\ifelse{html}{\out{atm-m<sup>3</sup>/L}}{\eqn{atm-m^3/mol}}) at 25 degrees Celsius (\ifelse{html}{\out{&degC}}{\eqn{^{\circ}C}})
}
\examples{
HrCalc(Hc = 0.0177)

}
