% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_importTemplateData.R
\name{importTemplateData}
\alias{importTemplateData}
\title{Import data using the vapintr data import template}
\usage{
importTemplateData(template_file_path)
}
\arguments{
\item{template_file_path}{Path to a copy of the vapintr data import template
with data to simulate}
}
\value{
A list with six elements, one for each type of parameter data
required to run the Johnson and Ettinger model and build output figures in
vapintr. The output of this function is the same as the output from
\code{processImportedTemplateData()}.
}
\description{
Reads in data from the vapintr Excel data import template and
passes the data to \code{processImportedTemplateData()}, which prepares the data
for simulation.
}
\examples{
#Get import data file path
import_data_file_path <- system.file("extdata",
"Stochastic_Simulation_Example_Data.xlsx", package = "vapintr")

#Import the file data
importTemplateData(import_data_file_path)

}
