% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_JEMParamDist.R
\name{JEMParamDist}
\alias{JEMParamDist}
\title{Define distributions for model input parameters}
\description{
An R6 class used to define input parameter distributions for
stochastic simulations. For more information on this class and how to use
it, see \code{vignette("StochasticSimulations", package = "vapintr")}.
}
\examples{
#Create an example JEMParamDist object for a parameter with a uniform distribution
JEMParamDist$new("Lb",
                 name = "Depth below grade to base of foundation",
                 units = "m",
                 dist_type = "Uniform",
                 minimum = 0.1,
                 maximum = 0.5)

#Create an example JEMParamDist object for a parameter with a truncated normal distribution
JEMParamDist$new("ach",
                 name = "Indoor air exchange rate",
                 units = "1/hr",
                 dist_type = "Truncated Normal",
                 minimum = 0.15,
                 arith_mean = 0.45,
                 arith_stdev = 0.1,
                 maximum = 1.26)


}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JEMParamDist-new}{\code{JEMParamDist$new()}}
\item \href{#method-JEMParamDist-getSymbol}{\code{JEMParamDist$getSymbol()}}
\item \href{#method-JEMParamDist-getDataFrameOfProperties}{\code{JEMParamDist$getDataFrameOfProperties()}}
\item \href{#method-JEMParamDist-getDistributionData}{\code{JEMParamDist$getDistributionData()}}
\item \href{#method-JEMParamDist-checkParamsAreNonNegative}{\code{JEMParamDist$checkParamsAreNonNegative()}}
\item \href{#method-JEMParamDist-paramsAreLessThanUpperBound}{\code{JEMParamDist$paramsAreLessThanUpperBound()}}
\item \href{#method-JEMParamDist-paramsAreGreaterThanLowerBound}{\code{JEMParamDist$paramsAreGreaterThanLowerBound()}}
\item \href{#method-JEMParamDist-getDistributionMinimum}{\code{JEMParamDist$getDistributionMinimum()}}
\item \href{#method-JEMParamDist-getDistributionMaximum}{\code{JEMParamDist$getDistributionMaximum()}}
\item \href{#method-JEMParamDist-clone}{\code{JEMParamDist$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JEMParamDist-new"></a>}}
\if{latex}{\out{\hypertarget{method-JEMParamDist-new}{}}}
\subsection{Method \code{new()}}{
Function to create a new \code{JEMParamDist} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JEMParamDist$new(
  symbol,
  name = NULL,
  units = NULL,
  dist_type = NULL,
  constant = NULL,
  minimum = NULL,
  mode = NULL,
  maximum = NULL,
  arith_mean = NULL,
  arith_stdev = NULL,
  geom_mean = NULL,
  geom_stdev = NULL,
  template_data_dfx = NULL,
  custom_distribution_data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Symbol used to define the parameter in the Johnson & Ettinger model}

\item{\code{name}}{Parameter name in the Johnson & Ettinger model or the contaminant name for concentration distributions}

\item{\code{units}}{Units associated with the distribution}

\item{\code{dist_type}}{Distribution type. Supported distributions are "Constant", "Uniform", "Triangular", "PERT", "Truncated Normal", "Truncated Lognormal", and "Custom"}

\item{\code{constant}}{Value of a constant distribution}

\item{\code{minimum}}{Distribution minimum}

\item{\code{mode}}{Distribution mode}

\item{\code{maximum}}{Distribution maximum}

\item{\code{arith_mean}}{Distribution arithmetic mean}

\item{\code{arith_stdev}}{Distribution standard deviation}

\item{\code{geom_mean}}{Distribution geometric mean}

\item{\code{geom_stdev}}{Distribution geometric standard deviation}

\item{\code{template_data_dfx}}{Data frame containing data imported from the vapintr data import template}

\item{\code{custom_distribution_data}}{Custom distribution data generated outside of the \code{JEMParamDist} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JEMParamDist-getSymbol"></a>}}
\if{latex}{\out{\hypertarget{method-JEMParamDist-getSymbol}{}}}
\subsection{Method \code{getSymbol()}}{
Function to return the value of the object's private "symbol" attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JEMParamDist$getSymbol()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JEMParamDist-getDataFrameOfProperties"></a>}}
\if{latex}{\out{\hypertarget{method-JEMParamDist-getDataFrameOfProperties}{}}}
\subsection{Method \code{getDataFrameOfProperties()}}{
Function to return a data frame of all the object's private attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JEMParamDist$getDataFrameOfProperties()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JEMParamDist-getDistributionData"></a>}}
\if{latex}{\out{\hypertarget{method-JEMParamDist-getDistributionData}{}}}
\subsection{Method \code{getDistributionData()}}{
Generate stochastic simulation data for the parameter based
on the object's assigned distribution and distribution parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JEMParamDist$getDistributionData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JEMParamDist-checkParamsAreNonNegative"></a>}}
\if{latex}{\out{\hypertarget{method-JEMParamDist-checkParamsAreNonNegative}{}}}
\subsection{Method \code{checkParamsAreNonNegative()}}{
Validate that all distribution parameter values are
non-negative. Distribution parameters assigned a value of zero are
allowed in this function. This function was not included as part of the
private validation function because values less than zero are
physically possible for some parameters (e.g., temperature).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JEMParamDist$checkParamsAreNonNegative()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JEMParamDist-paramsAreLessThanUpperBound"></a>}}
\if{latex}{\out{\hypertarget{method-JEMParamDist-paramsAreLessThanUpperBound}{}}}
\subsection{Method \code{paramsAreLessThanUpperBound()}}{
Validate that all distribution parameter values are less
than or equal to some upper limit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JEMParamDist$paramsAreLessThanUpperBound(upperBound, allowEqual = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{upperBound}}{The upper limit of the distribution parameter values}

\item{\code{allowEqual}}{A true/false flag identifying if the distribution
parameter values can equal the upper limit.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JEMParamDist-paramsAreGreaterThanLowerBound"></a>}}
\if{latex}{\out{\hypertarget{method-JEMParamDist-paramsAreGreaterThanLowerBound}{}}}
\subsection{Method \code{paramsAreGreaterThanLowerBound()}}{
Validate that all distribution parameter values are greater
than or equal to some lower limit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JEMParamDist$paramsAreGreaterThanLowerBound(lowerBound, allowEqual = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lowerBound}}{The lower limit of the distribution parameter values}

\item{\code{allowEqual}}{A true/false flag identifying if the distribution
parameter values can equal the lower limit}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JEMParamDist-getDistributionMinimum"></a>}}
\if{latex}{\out{\hypertarget{method-JEMParamDist-getDistributionMinimum}{}}}
\subsection{Method \code{getDistributionMinimum()}}{
Gets the minimum possible value for the parameter based
on the distribution type and distribution parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JEMParamDist$getDistributionMinimum()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JEMParamDist-getDistributionMaximum"></a>}}
\if{latex}{\out{\hypertarget{method-JEMParamDist-getDistributionMaximum}{}}}
\subsection{Method \code{getDistributionMaximum()}}{
Gets the maximum possible value for the parameter based
on the distribution type and distribution parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JEMParamDist$getDistributionMaximum()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JEMParamDist-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JEMParamDist-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JEMParamDist$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
