% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_MeasDataFig.R
\name{MeasDataFig}
\alias{MeasDataFig}
\title{Class for data in measured air concentration scatter plots}
\description{
An R6 Class used to build scatter plots of measured indoor and
outdoor air concentration data with lines for reference concentrations.
Most functions for building the plots are inherited from the parent
\code{ConcDataFig} class. This class just contains methods and attributes unique
to the measured data plots. See \code{vignette("ModelResultFigures", package = "vapintr")} for examples and more information on how this class is used
within vapintr.
}
\examples{
#Run the JEM using the example stochastic data
jem_results <- runTemplateData(stoc_jem_sim_example_data,
"Trichloroethylene")

#Create a new output figure from the JEM Monte Carlo results
#using the measured data in the example dataset
model_result_figure <- ModelResultFig$new(jem_results[[1]]$JEMResults,
"Trichloroethylene", "79-01-6", "ug/m3", "groundwater",
stoc_jem_sim_example_data[[1]])

#The object stored in the model_result_figure$meas_data_fig attribute is a
#MeasDataFig object
model_result_figure$meas_data_fig

}
\section{Super class}{
\code{\link[vapintr:ConcDataFig]{vapintr::ConcDataFig}} -> \code{MeasDataFig}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{measured_data_legend}}{A ggplot for the legend associated with the measured data}

\item{\code{reference_data_legend}}{A ggplot for the legend associated with the reference data}

\item{\code{measured_data}}{A data frame of measured data}

\item{\code{reference_data}}{A data frame of reference data}

\item{\code{x_axis_break_interval}}{The break interval for displaying date labels on the x-axis}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasDataFig-buildPlotElements}{\code{MeasDataFig$buildPlotElements()}}
\item \href{#method-MeasDataFig-setAxisLimits}{\code{MeasDataFig$setAxisLimits()}}
\item \href{#method-MeasDataFig-clone}{\code{MeasDataFig$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="addPlotLabels"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-addPlotLabels'><code>vapintr::ConcDataFig$addPlotLabels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="getStandardFactorColorFormats"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-getStandardFactorColorFormats'><code>vapintr::ConcDataFig$getStandardFactorColorFormats()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="getStandardFactorIndices"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-getStandardFactorIndices'><code>vapintr::ConcDataFig$getStandardFactorIndices()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="getStandardFactorLineFormats"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-getStandardFactorLineFormats'><code>vapintr::ConcDataFig$getStandardFactorLineFormats()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="getStandardFactorNames"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-getStandardFactorNames'><code>vapintr::ConcDataFig$getStandardFactorNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="getStandardFactorShapeFormats"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-getStandardFactorShapeFormats'><code>vapintr::ConcDataFig$getStandardFactorShapeFormats()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="getStandardFormatConstants"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-getStandardFormatConstants'><code>vapintr::ConcDataFig$getStandardFormatConstants()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="initialize"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-initialize'><code>vapintr::ConcDataFig$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasDataFig-buildPlotElements"></a>}}
\if{latex}{\out{\hypertarget{method-MeasDataFig-buildPlotElements}{}}}
\subsection{Method \code{buildPlotElements()}}{
Assembles all elements required to build the measured and reference data plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasDataFig$buildPlotElements()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasDataFig-setAxisLimits"></a>}}
\if{latex}{\out{\hypertarget{method-MeasDataFig-setAxisLimits}{}}}
\subsection{Method \code{setAxisLimits()}}{
Sets the x- and y-axis limits for the measured and reference data plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasDataFig$setAxisLimits(mod_data_fig = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod_data_fig}}{A \code{ModDataFig} object that will be displayed along with the measured and reference data in a combined figure.
Providing a value for this parameter will make both plots use the same y-axis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasDataFig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasDataFig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasDataFig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
