% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_ModDataFig.R
\name{ModDataFig}
\alias{ModDataFig}
\title{Class for data in modeled air concentration violin plots}
\description{
An R6 Class used to build violin plots of modeled data. Most
functions for building the plots are inherited from the parent \code{ConcDataFig}
class. This class just contains methods and attributes unique to the
modeled data violin plots. See \code{vignette("ModelResultFigures", package = "vapintr")} for examples and more information on how this class is used
within vapintr.
}
\examples{
#Run the JEM using the example stochastic data
jem_results <- runTemplateData(stoc_jem_sim_example_data, "Trichloroethylene")

#Create a new output figure from the JEM Monte Carlo results
model_result_figure <- ModelResultFig$new(jem_results[[1]]$JEMResults,
"Trichloroethylene", "79-01-6", "ug/m3", "groundwater")

#The object stored in the model_result_figure$mod_data_fig attribute is a
#ModDataFig object
model_result_figure$mod_data_fig

}
\section{Super class}{
\code{\link[vapintr:ConcDataFig]{vapintr::ConcDataFig}} -> \code{ModDataFig}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{violin_plot_legend}}{A ggplot for the legend associated with the violin plot data}

\item{\code{percentiles_legend}}{A ggplot for the legend associated with the violin plot percentiles}

\item{\code{modeled_data}}{A data frame of modeled indoor air concentration data}

\item{\code{percentile_50}}{The 50\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} percentile value of the modeled indoor air concentration data}

\item{\code{percentile_95}}{The 95\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} percentile value of the modeled indoor air concentration data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModDataFig-buildPlotElements}{\code{ModDataFig$buildPlotElements()}}
\item \href{#method-ModDataFig-setAxisLimits}{\code{ModDataFig$setAxisLimits()}}
\item \href{#method-ModDataFig-clone}{\code{ModDataFig$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="addPlotLabels"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-addPlotLabels'><code>vapintr::ConcDataFig$addPlotLabels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="getStandardFactorColorFormats"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-getStandardFactorColorFormats'><code>vapintr::ConcDataFig$getStandardFactorColorFormats()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="getStandardFactorIndices"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-getStandardFactorIndices'><code>vapintr::ConcDataFig$getStandardFactorIndices()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="getStandardFactorLineFormats"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-getStandardFactorLineFormats'><code>vapintr::ConcDataFig$getStandardFactorLineFormats()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="getStandardFactorNames"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-getStandardFactorNames'><code>vapintr::ConcDataFig$getStandardFactorNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="getStandardFactorShapeFormats"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-getStandardFactorShapeFormats'><code>vapintr::ConcDataFig$getStandardFactorShapeFormats()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="getStandardFormatConstants"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-getStandardFormatConstants'><code>vapintr::ConcDataFig$getStandardFormatConstants()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="ConcDataFig" data-id="initialize"><a href='../../vapintr/html/ConcDataFig.html#method-ConcDataFig-initialize'><code>vapintr::ConcDataFig$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModDataFig-buildPlotElements"></a>}}
\if{latex}{\out{\hypertarget{method-ModDataFig-buildPlotElements}{}}}
\subsection{Method \code{buildPlotElements()}}{
Assembles all elements required to build the modeled data violin plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModDataFig$buildPlotElements()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModDataFig-setAxisLimits"></a>}}
\if{latex}{\out{\hypertarget{method-ModDataFig-setAxisLimits}{}}}
\subsection{Method \code{setAxisLimits()}}{
Sets the y-axis limits for the modeled data violin plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModDataFig$setAxisLimits(meas_data_fig = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meas_data_fig}}{A \code{MeasDataFig} object that will be displayed along with the modeled data in a combined figure.
Providing a value for this parameter will make both plots use the same y-axis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModDataFig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModDataFig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModDataFig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
