% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_ModelResultFig.R
\name{ModelResultFig}
\alias{ModelResultFig}
\title{Create standard modeled and measured air concentration figures}
\description{
An R6 class used to build standard output figures from Johnson
and Ettinger model stochastic simulation results. See
\code{vignette("ModelResultFigures", package = "vapintr")} for examples and more
information on using this class.
}
\examples{
#Run the JEM using the example stochastic data
jem_results <- runTemplateData(stoc_jem_sim_example_data, "Trichloroethylene")

#Create a new output figure from the JEM Monte Carlo results
model_result_figure <- ModelResultFig$new(jem_results[[1]]$JEMResults,
"Trichloroethylene", "79-01-6", "ug/m3", "groundwater")

#Create the standard plot image from the model_result_figure object
model_result_figure$createStandardFigure()

}
\section{Super class}{
\code{\link[vapintr:StandardOutputFig]{vapintr::StandardOutputFig}} -> \code{ModelResultFig}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{contaminant_name}}{The name of the modeled contaminant}

\item{\code{contaminant_casrn}}{The contaminant's Chemical Abstracts Service Registry Number}

\item{\code{concentration_units}}{The units used to display concentrations on the y-axis ("ug/m3" or "ppb")}

\item{\code{source_medium}}{The contaminant source medium}

\item{\code{meas_data_fig}}{\code{MeasDataFig} object for the measured and reference data image (if included)}

\item{\code{mod_data_fig}}{\code{ModDataFig} object for the modeled data image}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelResultFig-new}{\code{ModelResultFig$new()}}
\item \href{#method-ModelResultFig-createStandardFigure}{\code{ModelResultFig$createStandardFigure()}}
\item \href{#method-ModelResultFig-buildPlotElements}{\code{ModelResultFig$buildPlotElements()}}
\item \href{#method-ModelResultFig-clone}{\code{ModelResultFig$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="StandardOutputFig" data-id="renderImageInNewWindow"><a href='../../vapintr/html/StandardOutputFig.html#method-StandardOutputFig-renderImageInNewWindow'><code>vapintr::StandardOutputFig$renderImageInNewWindow()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelResultFig-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelResultFig-new}{}}}
\subsection{Method \code{new()}}{
Function to create a new ModelResultFig object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelResultFig$new(
  jemMCResults,
  contaminant_name,
  contaminant_casrn,
  concentration_units,
  source_medium,
  measured_data_dfx = NULL,
  reference_data_dfx = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{jemMCResults}}{Monte Carlo simulation results from a stochastic simulation using \code{runJE()}}

\item{\code{contaminant_name}}{The name of the modeled contaminant for display in the figure images}

\item{\code{contaminant_casrn}}{The contaminant's Chemical Abstracts Service Registry Number}

\item{\code{concentration_units}}{The units used to display concentrations on the y-axis ("ug/m3" or "ppb")}

\item{\code{source_medium}}{The contaminant source medium}

\item{\code{measured_data_dfx}}{A data frame of measured data, typically imported from the vapintr data import template}

\item{\code{reference_data_dfx}}{A data frame of reference concentrations. Must include the fields "Name" and "Concentration".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelResultFig-createStandardFigure"></a>}}
\if{latex}{\out{\hypertarget{method-ModelResultFig-createStandardFigure}{}}}
\subsection{Method \code{createStandardFigure()}}{
Creates a standard figure using the modeled and measured data figures
assigned to the object's mod_data_fig and meas_data_fig attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelResultFig$createStandardFigure()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelResultFig-buildPlotElements"></a>}}
\if{latex}{\out{\hypertarget{method-ModelResultFig-buildPlotElements}{}}}
\subsection{Method \code{buildPlotElements()}}{
Sets the axes and calls the buildPlotElements() functions for the measured and modeled data figure objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelResultFig$buildPlotElements()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelResultFig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelResultFig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelResultFig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
