% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paramCalcs.R
\name{PercSatCalc}
\alias{PercSatCalc}
\title{Calculate the percent of pure component saturated vapor concentration (\%Sat)}
\usage{
PercSatCalc(Cs, Vc)
}
\arguments{
\item{Cs}{Source vapor concentration in micrograms per cubic meter (\ifelse{html}{\out{&microg/m<sup>3</sup>}}{\eqn{{\mu}g/m^3}})}

\item{Vc}{Saturated vapor concentration (\ifelse{html}{\out{&microg/m<sup>3</sup>}}{\eqn{{\mu}g/m^3}})}
}
\value{
The percent saturation. The value is returned as a decimal and not as
a percentage (i.e., a return value of 0.5 represents 50\%)
}
\description{
Calculates the percent of pure component saturated vapor concentration.
}
\examples{
PercSatCalc(Cs = 7238,
Vc = 1.65E8)

}
