% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_importTemplateData.R
\name{processImportedTemplateData}
\alias{processImportedTemplateData}
\title{Process imported data from the vapintr data import template}
\usage{
processImportedTemplateData(imported_data_lx)
}
\arguments{
\item{imported_data_lx}{A list with six elements, typically created within
\code{importTemplateData()}. In order, the six list elements represent data
read from the following sheets in the data import template:
\itemize{
\item ContaminantData
\item BuildingInfo
\item VadoseZoneInfo
\item StrataLogs
\item Settings
\item ReferenceAirConcentrations
}}
}
\value{
A list with six elements, one for each type of parameter data
required to run the Johnson and Ettinger model in vapintr. In order, the
list elements contain the following types of data:
\itemize{
\item Contaminant concentration data
\item Building parameter data
\item Vadoze zone parameter data
\item Soil strata data
\item Simulation settings
\item Reference air concentration data
}
}
\description{
Converts data imported from the vapintr data import template
into the form required by vapintr to run the Johnson and Ettinger model and
produce output figures. This function is not typically accessed directly,
but rather is called from \code{importTemplateData()}.
}
\examples{
#Get the data import template file path
import_data_file_path <- system.file("extdata",
"Stochastic_Simulation_Example_Data.xlsx", package = "vapintr")

#Import simulation settings
settings_dfx <- openxlsx::read.xlsx(import_data_file_path,
sheet = "Settings")

#Import building info
building_info_dfx <- openxlsx::read.xlsx(import_data_file_path,
sheet = "BuildingInfo", fillMergedCells = TRUE)

#Import vadose zone info
vadose_zone_info_dfx <- openxlsx::read.xlsx(import_data_file_path,
sheet = "VadoseZoneInfo", fillMergedCells = TRUE)

#Import soil strata log data
strata_logs_dfx <- openxlsx::read.xlsx(import_data_file_path,
sheet = "StrataLogs")

#Import contaminant data
contaminant_data_dfx <- openxlsx::read.xlsx(import_data_file_path,
sheet = "ContaminantData")

#Import reference air concentration data
reference_data_dfx <- openxlsx::read.xlsx(import_data_file_path,
sheet = "ReferenceAirConcentrations")

#Create list of data frames
imported_data_lx <- list (contaminant_data_dfx, building_info_dfx,
vadose_zone_info_dfx, strata_logs_dfx, settings_dfx, reference_data_dfx)

#Process the imported template data
processImportedTemplateData(imported_data_lx)

}
