% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pullResults.R
\name{pullResults}
\alias{pullResults}
\title{Parse output from \code{runJE()} into multiple thematic tables.}
\usage{
pullResults(simobj, i = NA)
}
\arguments{
\item{simobj}{A data frame or mc object output from \code{runJE()}.}

\item{i}{The Monte Carlo iteration number. Entering a value for i allows
for the results of a specific Monte Carlo iteration to be selected. If this
parameter is provided for a stochastic simulation, the tables produced are
the same as those produced for a deterministic simulation.}
}
\value{
A list of five data frames:
\itemize{
\item "source" describes source characteristics
\item "building" describes building characteristics
\item "subsurface" describes intermediate subsurface calculations
\item "abc_parameter" describes the Johnson and Ettinger model A, B, and C
parameters
\item "output" identifies the attenuation factor (alpha) and the final
concentration outputs
}
}
\description{
Takes output from \code{runJE()} and provides a list of
five tables that summarize the inputs, intermediate calculations, and
output of the Johnson and Ettinger model.
}
\examples{
#Get source data
contaminant_data <- det_jem_sim_example_data[[1]]
building_data <- det_jem_sim_example_data[[2]]
vadose_zone_data <- det_jem_sim_example_data[[3]]
strata_logs_data <- det_jem_sim_example_data[[4]]
settings_data <- det_jem_sim_example_data[[5]]

#Run the Johnson and Ettinger Model
runJE_output <- runJE(contaminant_data, building_data, vadose_zone_data,
strata_logs_data, settings_data)

#Get the output tables
pullResults(runJE_output)

}
