% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paramCalcs.R
\name{QbCalc}
\alias{QbCalc}
\title{Calculate the building ventilation rate (Qb)}
\usage{
QbCalc(Abf, Hb, ach)
}
\arguments{
\item{Abf}{Enclosed space floor area in square meters (\ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^2}})}

\item{Hb}{Enclosed space mixing height in meters (\ifelse{html}{\out{m}}{\eqn{m}})}

\item{ach}{Indoor air exchange rate in air changes per hour (\ifelse{html}{\out{hr<sup>-1</sup>}}{\eqn{hr^-1}})}
}
\value{
The building ventilation rate (\ifelse{html}{\out{m<sup>3</sup>/hr}}{\eqn{m^3/hr}})
}
\description{
Calculates the building ventilation rate in cubic meters per
hour (\ifelse{html}{\out{m<sup>3</sup>/hr}}{\eqn{m^3/hr}})
}
\examples{
QbCalc(Abf = 150,
Hb = 2.44,
ach = 0.45)

}
