% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paramCalcs.R
\name{QsoilCalc}
\alias{QsoilCalc}
\title{Calculate the average vapor flow rate into the building (Qsoil)}
\usage{
QsoilCalc(Qsoil_Qb, Qb, foundation_type)
}
\arguments{
\item{Qsoil_Qb}{The ratio of the average vapor flow rate into the building
divided by the building ventilation rate (dimensionless)}

\item{Qb}{Building ventilation rate (\ifelse{html}{\out{m<sup>3</sup>/hr}}{\eqn{m^3/hr}})}

\item{foundation_type}{The foundation type. Choose from the following:
\itemize{
\item "Slab-grade" for a slab-on-grade foundation
\item "Basement-slab" for a basement with slab foundation
\item "Basement-dirt" for a basement with dirt floor foundation
\item "Crawlspace-slab" for a crawlspace with slab foundation
\item "Crawlspace-dirt" for a crawlspace with dirt floor foundation
}}
}
\value{
The average vapor flow rate into the building (\ifelse{html}{\out{m<sup>3</sup>/hr}}{\eqn{m^3/hr}})
}
\description{
Calculates the average vapor flow rate into the building in
cubic meters per hour (\ifelse{html}{\out{m<sup>3</sup>/hr}}{\eqn{m^3/hr}})
}
\examples{
QsoilCalc(Qsoil_Qb = 0.0030,
Qb = 164.70,
foundation_type = "Slab-grade")

}
