% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_runTemplateData.R
\name{runTemplateData}
\alias{runTemplateData}
\title{Run a Johnson and Ettinger model simulation using the vapintr data
import template}
\usage{
runTemplateData(
  template_data,
  contaminant = NA,
  logID = NA,
  use_aggregate_data_in_det_simulation = FALSE,
  capture_warnings_and_errors = FALSE
)
}
\arguments{
\item{template_data}{A list of data frames with parameter inputs for the
Johnson and Ettinger model. This list can be produced using
\code{importTemplateData()} and is the same as the output from that function.}

\item{contaminant}{An optional parameter identifying the name of the
contaminant to be simulated from the imported data. The default value is
\code{NA}. If a contaminant name is not provided, the function will attempt to
run the JEM for all contaminants in the imported template data.}

\item{logID}{The LogID of a soil strata log from the imported data set to use
in the simulation. This parameter is optional and has a default value of
\code{NA}. If it is left as \code{NA} in a deterministic simulation, the simulation
will use the first LogID identified in the soil strata log data frame
imported from the template. If \code{NA} in a stochastic simulation, the
simulation will randomly sample from the supplied LogIDs for each Monte
Carlo iteration.}

\item{use_aggregate_data_in_det_simulation}{True/False parameter for
deterministic simulations that indicates whether to aggregate the data by
contaminant and sample location ID (True) or to keep contaminant records
separate (False). See \code{vignette("DeterministicSimulations", package = "vapintr")} for more information.}

\item{capture_warnings_and_errors}{True/False parameter indicating whether to
capture warnings and errors associated with individual simulations or write
them to the console. Default is FALSE.}
}
\value{
A named output list with information to identify each JEM simulation
run. In both deterministic and stochastic simulations, the JEM results from
\code{runJE()} are stored in the \code{JEMResults} field of the output list elements,
and the \code{Error} and \code{Warning} fields identify any errors or warnings thrown
during each simulation. The other fields in each list element depend on the
type of simulation performed and identify inputs associated with the
results.
}
\description{
This function is a wrapper function for \code{runJE()} that supports
simulations using vapintr's data import template. See the following
vignettes for more information.

\code{vignette("DeterministicSimulations", package = "vapintr")}

\code{vignette("StochasticSimulations", package = "vapintr")}
}
\examples{

#Run a deterministic simulation
#Get import data file path
import_data_file_path <- system.file("extdata",
"Deterministic_Simulation_Example_Data.xlsx", package = "vapintr")

#Import the file data
template_data <- importTemplateData(import_data_file_path)

#Run the simulation
runTemplateData(template_data)

#Run a stochastic simulation
import_data_file_path <- system.file("extdata",
"Stochastic_Simulation_Example_Data.xlsx", package = "vapintr")

#Import the file data
template_data <- importTemplateData(import_data_file_path)

#Run the simulation
runTemplateData(template_data)

}
