% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SCS_soil}
\alias{SCS_soil}
\title{Soil Conservation Service soil property data}
\format{
A data frame with the following fields:
\describe{
\item{SCS_Type}{SCS soil type name}
\item{n}{total porosity (dimensionless)}
\item{nw}{water-filled porosity (dimensionless)}
\item{rho}{bulk density (grams per cubic centimeter, or \ifelse{html}{\out{g/cm<sup>2</sup>}}{\eqn{g/cm^2}})}
\item{ncz}{capillary zone water filled porosity (dimensionless)}
\item{hcz}{capillary zone height (centimeters, or \ifelse{html}{\out{cm}}{\eqn{cm}})}
}
}
\usage{
SCS_soil
}
\description{
A data frame of properties for the 12 Soil Conservation Service
(SCS) soil types. All properties were imported from version 6.0 of USEPA's
Johnson and Ettinger model spreadsheet.
}
\keyword{datasets}
