% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_SiteModelFig.R
\name{SiteModelFig}
\alias{SiteModelFig}
\title{Create conceptual site model figures}
\description{
An R6 Class used to build standard conceptual site model figures
based on the settings data inputs for \code{runJE()}. For more information on
this class and how to use it, see the following vignette.

\code{vignette("ConceptualSiteModelAndSoilStrataFigures", package = "vapintr")}
}
\examples{
#Get settings data from the example data
settings_data <- det_jem_sim_example_data[[5]]

#Initialize the site model figure object
site_model_figure <- SiteModelFig$new(settings_data)

#Create the standard image using the SiteModelFig object
site_model_figure$createStandardFigure()

#Plot the standard image
plot(site_model_figure$fig_image)

}
\section{Super class}{
\code{\link[vapintr:StandardOutputFig]{vapintr::StandardOutputFig}} -> \code{SiteModelFig}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{settings_data_dfx}}{Data frame of simulation settings data. Should
have the same format as the settings_data parameter from \code{runJE()}.}

\item{\code{site_model_fig}}{Site model image for plotting}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SiteModelFig-new}{\code{SiteModelFig$new()}}
\item \href{#method-SiteModelFig-createStandardFigure}{\code{SiteModelFig$createStandardFigure()}}
\item \href{#method-SiteModelFig-clone}{\code{SiteModelFig$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="StandardOutputFig" data-id="renderImageInNewWindow"><a href='../../vapintr/html/StandardOutputFig.html#method-StandardOutputFig-renderImageInNewWindow'><code>vapintr::StandardOutputFig$renderImageInNewWindow()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteModelFig-new"></a>}}
\if{latex}{\out{\hypertarget{method-SiteModelFig-new}{}}}
\subsection{Method \code{new()}}{
Function to create a new \code{SiteModelFig} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteModelFig$new(settings_data_dfx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings_data_dfx}}{Data frame of simulation settings data. Should
have the same format as the settings_data parameter from \code{runJE()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteModelFig-createStandardFigure"></a>}}
\if{latex}{\out{\hypertarget{method-SiteModelFig-createStandardFigure}{}}}
\subsection{Method \code{createStandardFigure()}}{
Assigns values to the standard properties inherited from the
\code{StandardOutputFig} class used to render the figure at a recommended size.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteModelFig$createStandardFigure()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiteModelFig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SiteModelFig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiteModelFig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
