% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_StandardOutputFig.R
\name{StandardOutputFig}
\alias{StandardOutputFig}
\title{Parent class for creating standard output figures}
\description{
A parent R6 Class used to build standard vapintr output figures
at recommended sizes. It has three properties: one for the image to be
rendered (fig_image), one for the recommended width (fig_width), and one
for the recommended height (fig_height). The object's
renderImageInNewWindow() function will render the image saved in the
fig_image property in a window at the dimensions specified in the object's
fig_width and fig_height properties.
}
\examples{
StandardOutputFig$new()

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fig_image}}{The figure image to be rendered}

\item{\code{fig_width}}{The recommended width of the figure, in inches}

\item{\code{fig_height}}{The recommended height of the figure, in inches}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StandardOutputFig-renderImageInNewWindow}{\code{StandardOutputFig$renderImageInNewWindow()}}
\item \href{#method-StandardOutputFig-clone}{\code{StandardOutputFig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardOutputFig-renderImageInNewWindow"></a>}}
\if{latex}{\out{\hypertarget{method-StandardOutputFig-renderImageInNewWindow}{}}}
\subsection{Method \code{renderImageInNewWindow()}}{
Helper function for creating images in a new window of the
recommended size in RStudio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardOutputFig$renderImageInNewWindow()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StandardOutputFig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StandardOutputFig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandardOutputFig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
