% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stoc_jem_sim_example_data}
\alias{stoc_jem_sim_example_data}
\title{Stochastic simulation data for running function examples}
\format{
An object of class \code{list} of length 6.
}
\usage{
stoc_jem_sim_example_data
}
\description{
A list of 6 elements used for stochastic simulation function
examples. The first element contains a table of concentration data that
must be converted to a \code{JEMParamDist} object using \code{getInputConc()} before
entering it in \code{runJE()}. The following function call converts the
concentration data into the form required by the contaminant_data parameter
in \code{runJE()}:

\code{getInputConc(stoc_jem_sim_example_data[[1]], "Trichloroethylene", "Groundwater", "MC")}.

The next four list elements can be input directly into \code{runJE()} and
correspond with the following \code{runJE()} inputs. The 6th element is a data
frame of reference air concentration data which can be loaded into result
figures but is not required for \code{runJE()}.
\enumerate{
\item building_data
\item vadose_zone_data
\item strata_logs_data
\item settings_data
}
}
\keyword{datasets}
