% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_StrataProfileFig.R
\name{StrataProfileFig}
\alias{StrataProfileFig}
\title{Create soil strata log figures}
\description{
An R6 Class used to build standard soil strata log figures
based on the soil strata inputs to \code{runJE()}. For more information on
this class and how to use it, see the following vignette.

\code{vignette("ConceptualSiteModelAndSoilStrataFigures", package = "vapintr")}
}
\examples{
#Get soil strata data from the example data
strata_logs_data <- det_jem_sim_example_data[[4]]

#Initialize the soil strata logs figure object
strata_logs_figure <- StrataProfileFig$new(strata_logs_data)

#Create the standard image using the StrataProfileFig object
strata_logs_figure$createStandardFigure()

#Plot the standard image
plot(strata_logs_figure$fig_image)
}
\section{Super class}{
\code{\link[vapintr:StandardOutputFig]{vapintr::StandardOutputFig}} -> \code{StrataProfileFig}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{strata_logs_dfx}}{Data frame of strata log data. Should have the
same format as the strata_logs_data input parameter to \code{runJE()}.}

\item{\code{strata_log_figs_lx}}{List of ggplot strata log figures produced
by the \code{StrataProfileFig} object}

\item{\code{maximum_depth}}{Maximum depth of any of the strata logs in the figure}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StrataProfileFig-new}{\code{StrataProfileFig$new()}}
\item \href{#method-StrataProfileFig-createStandardFigure}{\code{StrataProfileFig$createStandardFigure()}}
\item \href{#method-StrataProfileFig-clone}{\code{StrataProfileFig$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vapintr" data-topic="StandardOutputFig" data-id="renderImageInNewWindow"><a href='../../vapintr/html/StandardOutputFig.html#method-StandardOutputFig-renderImageInNewWindow'><code>vapintr::StandardOutputFig$renderImageInNewWindow()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrataProfileFig-new"></a>}}
\if{latex}{\out{\hypertarget{method-StrataProfileFig-new}{}}}
\subsection{Method \code{new()}}{
Function to create a new \code{StrataProfileFig} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrataProfileFig$new(strata_logs_dfx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{strata_logs_dfx}}{Data frame of strata log data. Should have the
same format as the strata_logs_data input parameter to \code{runJE()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrataProfileFig-createStandardFigure"></a>}}
\if{latex}{\out{\hypertarget{method-StrataProfileFig-createStandardFigure}{}}}
\subsection{Method \code{createStandardFigure()}}{
Assigns values to the standard properties inherited from the
\code{StandardOutputFig} class used to render the figure at a recommended size.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrataProfileFig$createStandardFigure(plot_orientation = "default")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot_orientation}}{Controls whether the output figure will have a
"portrait", "landscape", or "default" orientation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrataProfileFig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StrataProfileFig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrataProfileFig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
