% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapintr_package.R
\docType{package}
\name{vapintr-package}
\alias{vapintr}
\alias{vapintr-package}
\title{vapintr: Run the Johnson and Ettinger Groundwater Vapor Intrusion model in R}
\description{
This package runs the USEPA Johnson and Ettinger model for simulating vapor intrusion in R, allowing for Monte Carlo simulation of uncertain parameters.
}
\author{
\strong{Maintainer}: James Durant \email{hzd3@cdc.gov} (\href{https://orcid.org/0000-0002-3378-7386}{ORCID})

Authors:
\itemize{
  \item William Morgan \email{William.Morgan@erg.com} (\href{https://orcid.org/0009-0008-9931-2237}{ORCID})
}

Other contributors:
\itemize{
  \item Kirsten Miller \email{Kirsten.Miller@erg.com} [contributor]
  \item Tonia Burke \email{fxt9@cdc.gov} [reviewer]
  \item Sandra Miller \email{nyz7@cdc.gov} [reviewer]
  \item Tebitha Mawokomatanda \email{chn4@cdc.gov} [reviewer]
}

}
\keyword{internal}
